% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_schema.r
\name{enforce_schema}
\alias{enforce_schema}
\alias{enforce_schema.with_schema}
\title{Enforce an attached schema}
\usage{
enforce_schema(.data, ...)

\method{enforce_schema}{with_schema}(
  .data,
  ...,
  .error_call = caller_env(),
  .darg = caller_arg(.data)
)
}
\arguments{
\item{.data}{a data.frame or list containing an attached schema (of
class \code{with_schema}).}

\item{...}{for future extensions - leave blank.}

\item{.error_call}{the call environment to use for error messages
(passed to \link[rlang:abort]{rlang::abort}).}

\item{.darg}{the argument name of \code{.data} to use in error messages.}
}
\value{
Object \code{.data}, with no change if attached schema is from \code{schema},
or with named elements cast or recycled to the desired type/size if
attached schema is from \code{schema_cast} or \code{schema_recycle} respectively.
}
\description{
For objects with an attached schema (from \link{schema}, \link{schema_cast}, or
\link{schema_recycle}), enforce the schema on the object.
}
\details{
See \link{schema}, \link{schema_cast}
and \link{schema_recycle} for creating and attaching a schema.
}
\examples{
# NB: Some of these examples are expected to produce an error. To
#     prevent them from terminating a run with example() they are
#     piped into a call to try().

li <- list(x = 1, y = "hi")
li_with_schema <- schema(li, x == 1, is.character(y))

li_with_schema$y <- 1
enforce_schema(li_with_schema) |> try()
# Error:
# Caused by error in `enforce_schema()`:
# ℹ In argument: `is.character(y)` for data mask `li_with_schema`.
# ! Returned `FALSE`.

df <- data.frame(x = 1:2)
df_with_schema <- schema_cast(df, x = integer(), .lossy = TRUE)

df_with_schema$x <- c(1.5, 2.5)
enforce_schema(df_with_schema)$x
# 1 2

li_with_schema <- schema_recycle(li, x = 2, y = 3)
li_with_schema$y <- "hi"
enforce_schema(li_with_schema)$y
# "hi" "hi" "hi"

# The `.error_call` argument can be used to specify where the error occurs,
# by default this is the caller environment.
myfunc <- function(.x) enforce_schema(.x)
li_with_schema$x <- 1:3
myfunc(li_with_schema) |> try()
# Error in `myfunc()`:
# Caused by error in `enforce_schema()`:
# ℹ In argument: `x = 2` for data mask `li`.
# ! Can't recycle `x` (size 3) to size 2.
}
