% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{require_col_one_of}
\alias{require_col_one_of}
\title{Require Column Values from a Set}
\usage{
require_col_one_of(restriction, col, values)
}
\arguments{
\item{restriction}{a \code{restriction} object.}

\item{col}{character(1) column name.}

\item{values}{vector of allowed values.}
}
\value{
The modified \code{restriction} object.
}
\description{
Validates that all values in a column are among the allowed values.
}
\seealso{
Other column checks: 
\code{\link{require_col_between}()},
\code{\link{require_col_character}()},
\code{\link{require_col_numeric}()}
}
\concept{column checks}
