% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require-sets.R
\name{require_contains}
\alias{require_contains}
\title{Require All of a Set of Values}
\usage{
require_contains(restriction, values)
}
\arguments{
\item{restriction}{a \code{restriction} object.}

\item{values}{vector of values that must be present.}
}
\value{
The modified \code{restriction} object.
}
\description{
Validates that the value contains every element of \code{values}. Extra
elements are allowed.
}
\details{
\code{NA} elements of the validated value are ignored. A factor is
compared by its labels. For the opposite direction (every element must be
one of \code{values}) use \code{\link[=require_one_of]{require_one_of()}}.
}
\seealso{
Other value checks:
\code{\link[=require_between]{require_between()}},
\code{\link[=require_disjoint]{require_disjoint()}},
\code{\link[=require_levels]{require_levels()}},
\code{\link[=require_negative]{require_negative()}},
\code{\link[=require_one_of]{require_one_of()}},
\code{\link[=require_positive]{require_positive()}},
\code{\link[=require_set_equal]{require_set_equal()}},
\code{\link[=require_unique]{require_unique()}}
}
\concept{value checks}
