% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{require_df}
\alias{require_df}
\title{Require a Data Frame}
\usage{
require_df(restriction)
}
\arguments{
\item{restriction}{a \code{restriction} object.}
}
\value{
The modified \code{restriction} object.
}
\description{
Validates that the value is a \code{data.frame}.
}
