% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{require_dim}
\alias{require_dim}
\title{Require Exact Dimensions}
\usage{
require_dim(restriction, dims)
}
\arguments{
\item{restriction}{a \code{restriction} object.}

\item{dims}{numeric vector of required extents, one per dimension; \code{NA}
leaves that dimension unchecked.}
}
\value{
The modified \code{restriction} object.
}
\description{
Validates \code{dim(value)} of a data.frame, matrix or array. \code{NA} entries in
\code{dims} match any extent.
}
\examples{
design <- restrict("X") |> require_class("matrix") |> require_dim(c(NA, 3))
design(matrix(0, 10, 3))
try(design(matrix(0, 10, 2)))

}
\seealso{
Other structure checks:
\code{\link[=require_has_cols]{require_has_cols()}},
\code{\link[=require_length]{require_length()}},
\code{\link[=require_length_matches]{require_length_matches()}},
\code{\link[=require_length_max]{require_length_max()}},
\code{\link[=require_length_min]{require_length_min()}},
\code{\link[=require_named]{require_named()}},
\code{\link[=require_names]{require_names()}},
\code{\link[=require_ncol_matches]{require_ncol_matches()}},
\code{\link[=require_ncol_min]{require_ncol_min()}},
\code{\link[=require_nrow_matches]{require_nrow_matches()}},
\code{\link[=require_nrow_min]{require_nrow_min()}},
\code{\link[=require_scalar]{require_scalar()}},
\code{\link[=require_sorted]{require_sorted()}},
\code{\link[=require_unique_names]{require_unique_names()}}
}
\concept{structure checks}
