% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require-sets.R
\name{require_has_cols}
\alias{require_has_cols}
\title{Require Specific Columns}
\usage{
require_has_cols(restriction, cols)
}
\arguments{
\item{restriction}{a \code{restriction} object.}

\item{cols}{character vector of required column names.}
}
\value{
The modified \code{restriction} object.
}
\description{
Validates that a data.frame contains all specified columns. Equivalent to
\code{require_names(cols, mode = "superset")} with column wording in the
message.
}
\seealso{
Other structure checks:
\code{\link[=require_dim]{require_dim()}},
\code{\link[=require_length]{require_length()}},
\code{\link[=require_length_matches]{require_length_matches()}},
\code{\link[=require_length_max]{require_length_max()}},
\code{\link[=require_length_min]{require_length_min()}},
\code{\link[=require_named]{require_named()}},
\code{\link[=require_names]{require_names()}},
\code{\link[=require_ncol_matches]{require_ncol_matches()}},
\code{\link[=require_ncol_min]{require_ncol_min()}},
\code{\link[=require_nrow_matches]{require_nrow_matches()}},
\code{\link[=require_nrow_min]{require_nrow_min()}},
\code{\link[=require_scalar]{require_scalar()}},
\code{\link[=require_sorted]{require_sorted()}},
\code{\link[=require_unique_names]{require_unique_names()}}
}
\concept{structure checks}
