% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{require_length_max}
\alias{require_length_max}
\title{Require Maximum Length}
\usage{
require_length_max(restriction, n)
}
\arguments{
\item{restriction}{a \code{restriction} object.}

\item{n}{integer(1) maximum length.}
}
\value{
The modified \code{restriction} object.
}
\description{
Validates that the value has at most length \code{n}.
}
\seealso{
Other structure checks:
\code{\link[=require_dim]{require_dim()}},
\code{\link[=require_has_cols]{require_has_cols()}},
\code{\link[=require_length]{require_length()}},
\code{\link[=require_length_matches]{require_length_matches()}},
\code{\link[=require_length_min]{require_length_min()}},
\code{\link[=require_named]{require_named()}},
\code{\link[=require_names]{require_names()}},
\code{\link[=require_ncol_matches]{require_ncol_matches()}},
\code{\link[=require_ncol_min]{require_ncol_min()}},
\code{\link[=require_nrow_matches]{require_nrow_matches()}},
\code{\link[=require_nrow_min]{require_nrow_min()}},
\code{\link[=require_scalar]{require_scalar()}},
\code{\link[=require_sorted]{require_sorted()}},
\code{\link[=require_unique_names]{require_unique_names()}}
}
\concept{structure checks}
