% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{require_logical}
\alias{require_logical}
\title{Require Logical Type}
\usage{
require_logical(restriction, no_na = FALSE)
}
\arguments{
\item{restriction}{a \code{restriction} object.}

\item{no_na}{logical; if \code{TRUE}, rejects NA values.}
}
\value{
The modified \code{restriction} object.
}
\description{
Validates that the value is logical. Optionally checks for NA values.
}
\seealso{
Other type checks: 
\code{\link{require_character}()},
\code{\link{require_df}()},
\code{\link{require_integer}()},
\code{\link{require_numeric}()}
}
\concept{type checks}
