% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{require_no_na}
\alias{require_no_na}
\title{Require No NA Values}
\usage{
require_no_na(restriction)
}
\arguments{
\item{restriction}{a \code{restriction} object.}
}
\value{
The modified \code{restriction} object.
}
\description{
Validates that the value contains no \code{NA} values. Works on any atomic type.
}
\seealso{
Other missingness checks: 
\code{\link{require_finite}()}
}
\concept{missingness checks}
