% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{require_not_null}
\alias{require_not_null}
\title{Require Non-NULL Value}
\usage{
require_not_null(restriction)
}
\arguments{
\item{restriction}{a \code{restriction} object.}
}
\value{
The modified \code{restriction} object.
}
\description{
Validates that the value is not \code{NULL}. Place this step first in the
pipeline when \code{NULL} is a possible input.
}
\seealso{
Other missingness checks: 
\code{\link{require_finite}()},
\code{\link{require_no_na}()}
}
\concept{missingness checks}
