% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require-sets.R
\name{require_set_equal}
\alias{require_set_equal}
\title{Require the Same Set of Values}
\usage{
require_set_equal(restriction, values)
}
\arguments{
\item{restriction}{a \code{restriction} object.}

\item{values}{vector of values that must be present.}
}
\value{
The modified \code{restriction} object.
}
\description{
Validates that the value and \code{values} hold the same distinct elements.
Order and duplicates are ignored.
}
\seealso{
Other value checks:
\code{\link[=require_between]{require_between()}},
\code{\link[=require_contains]{require_contains()}},
\code{\link[=require_disjoint]{require_disjoint()}},
\code{\link[=require_levels]{require_levels()}},
\code{\link[=require_negative]{require_negative()}},
\code{\link[=require_one_of]{require_one_of()}},
\code{\link[=require_positive]{require_positive()}},
\code{\link[=require_unique]{require_unique()}}
}
\concept{value checks}
