% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{require_unique}
\alias{require_unique}
\title{Require Unique Values}
\usage{
require_unique(restriction)
}
\arguments{
\item{restriction}{a \code{restriction} object.}
}
\value{
The modified \code{restriction} object.
}
\description{
Validates that the value contains no duplicates. Reports the positions
of duplicated elements.
}
\seealso{
Other value checks:
\code{\link[=require_between]{require_between()}},
\code{\link[=require_contains]{require_contains()}},
\code{\link[=require_disjoint]{require_disjoint()}},
\code{\link[=require_levels]{require_levels()}},
\code{\link[=require_negative]{require_negative()}},
\code{\link[=require_one_of]{require_one_of()}},
\code{\link[=require_positive]{require_positive()}},
\code{\link[=require_set_equal]{require_set_equal()}}
}
\concept{value checks}
