% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-package.R
\docType{package}
\name{restrictR-package}
\alias{restrictR}
\alias{restrictR-package}
\title{restrictR: Composable Runtime Contracts for R}
\description{
Build reusable validators from small building blocks using the base pipe operator. Define runtime contracts once with 'restrict()' and enforce them anywhere in code. Validators compose naturally, support dependent rules via formulas, and produce clear, path-aware error messages. No DSL, no operator overloading, just idiomatic R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gillescolling.com/restrictR/}
  \item \url{https://github.com/gcol33/restrictR}
  \item Report bugs at \url{https://github.com/gcol33/restrictR/issues}
}

}
\author{
\strong{Maintainer}: Gilles Colling \email{gilles.colling051@gmail.com} (\href{https://orcid.org/0000-0003-3070-6066}{ORCID}) [copyright holder]

}
\keyword{internal}
