% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{require_between}
\alias{require_between}
\title{Require Value in Range}
\usage{
require_between(
  restriction,
  lower = -Inf,
  upper = Inf,
  exclusive_lower = FALSE,
  exclusive_upper = FALSE
)
}
\arguments{
\item{restriction}{a \code{restriction} object.}

\item{lower}{numeric(1) lower bound (default \code{-Inf}).}

\item{upper}{numeric(1) upper bound (default \code{Inf}).}

\item{exclusive_lower}{logical; if \code{TRUE}, lower bound is exclusive.}

\item{exclusive_upper}{logical; if \code{TRUE}, upper bound is exclusive.}
}
\value{
The modified \code{restriction} object.
}
\description{
Validates that all elements of a numeric value fall within a specified range.
}
\seealso{
Other value checks: 
\code{\link{require_negative}()},
\code{\link{require_one_of}()},
\code{\link{require_positive}()}
}
\concept{value checks}
