% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{require_character}
\alias{require_character}
\title{Require Character Type}
\usage{
require_character(restriction, no_na = FALSE)
}
\arguments{
\item{restriction}{a \code{restriction} object.}

\item{no_na}{logical; if \code{TRUE}, rejects NA values.}
}
\value{
The modified \code{restriction} object.
}
\description{
Validates that the value is character. Optionally checks for NA values.
}
\seealso{
Other type checks: 
\code{\link{require_df}()},
\code{\link{require_numeric}()}
}
\concept{type checks}
