% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{require_has_cols}
\alias{require_has_cols}
\title{Require Specific Columns}
\usage{
require_has_cols(restriction, cols)
}
\arguments{
\item{restriction}{a \code{restriction} object.}

\item{cols}{character vector of required column names.}
}
\value{
The modified \code{restriction} object.
}
\description{
Validates that a data.frame contains all specified columns.
}
\seealso{
Other structure checks: 
\code{\link{require_length}()},
\code{\link{require_length_matches}()},
\code{\link{require_length_max}()},
\code{\link{require_length_min}()},
\code{\link{require_nrow_matches}()},
\code{\link{require_nrow_min}()}
}
\concept{structure checks}
