% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{require_length_min}
\alias{require_length_min}
\title{Require Minimum Length}
\usage{
require_length_min(restriction, n)
}
\arguments{
\item{restriction}{a \code{restriction} object.}

\item{n}{integer(1) minimum length.}
}
\value{
The modified \code{restriction} object.
}
\description{
Validates that the value has at least length \code{n}.
}
\seealso{
Other structure checks: 
\code{\link{require_has_cols}()},
\code{\link{require_length}()},
\code{\link{require_length_matches}()},
\code{\link{require_length_max}()},
\code{\link{require_named}()},
\code{\link{require_nrow_matches}()},
\code{\link{require_nrow_min}()},
\code{\link{require_scalar}()}
}
\concept{structure checks}
