% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{require_named}
\alias{require_named}
\title{Require Named Value}
\usage{
require_named(restriction)
}
\arguments{
\item{restriction}{a \code{restriction} object.}
}
\value{
The modified \code{restriction} object.
}
\description{
Validates that the value has names. Useful for named vectors and lists.
}
\seealso{
Other structure checks: 
\code{\link{require_has_cols}()},
\code{\link{require_length}()},
\code{\link{require_length_matches}()},
\code{\link{require_length_max}()},
\code{\link{require_length_min}()},
\code{\link{require_nrow_matches}()},
\code{\link{require_nrow_min}()},
\code{\link{require_scalar}()}
}
\concept{structure checks}
