% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{require_nrow_min}
\alias{require_nrow_min}
\title{Require Minimum Number of Rows}
\usage{
require_nrow_min(restriction, n)
}
\arguments{
\item{restriction}{a \code{restriction} object.}

\item{n}{integer(1) minimum row count.}
}
\value{
The modified \code{restriction} object.
}
\description{
Validates that a data.frame has at least \code{n} rows.
}
