% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{require_one_of}
\alias{require_one_of}
\title{Require Value from a Set}
\usage{
require_one_of(restriction, values)
}
\arguments{
\item{restriction}{a \code{restriction} object.}

\item{values}{vector of allowed values.}
}
\value{
The modified \code{restriction} object.
}
\description{
Validates that the value is one of the allowed values.
}
