% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{require_one_of}
\alias{require_one_of}
\title{Require Value from a Set}
\usage{
require_one_of(restriction, values)
}
\arguments{
\item{restriction}{a \code{restriction} object.}

\item{values}{vector of allowed values.}
}
\value{
The modified \code{restriction} object.
}
\description{
Validates that all elements of the value are among the allowed values.
}
\seealso{
Other value checks: 
\code{\link{require_between}()},
\code{\link{require_negative}()},
\code{\link{require_positive}()}
}
\concept{value checks}
