% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{require_scalar}
\alias{require_scalar}
\title{Require Scalar Value}
\usage{
require_scalar(restriction)
}
\arguments{
\item{restriction}{a \code{restriction} object.}
}
\value{
The modified \code{restriction} object.
}
\description{
Validates that the value has length 1. Rejects \code{NULL}, zero-length vectors,
and vectors with more than one element.
}
\seealso{
Other structure checks: 
\code{\link{require_has_cols}()},
\code{\link{require_length}()},
\code{\link{require_length_matches}()},
\code{\link{require_length_max}()},
\code{\link{require_length_min}()},
\code{\link{require_named}()},
\code{\link{require_nrow_matches}()},
\code{\link{require_nrow_min}()}
}
\concept{structure checks}
