% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{require_unique}
\alias{require_unique}
\title{Require Unique Values}
\usage{
require_unique(restriction)
}
\arguments{
\item{restriction}{a \code{restriction} object.}
}
\value{
The modified \code{restriction} object.
}
\description{
Validates that the value contains no duplicates. Reports the positions
of duplicated elements.
}
\seealso{
Other value checks: 
\code{\link{require_between}()},
\code{\link{require_negative}()},
\code{\link{require_one_of}()},
\code{\link{require_positive}()}
}
\concept{value checks}
